+++ src/avr/avr_timer0.h
@@ -98,7 +98,7 @@ public:
     // La función clock_period_in_us traduce la forma de hablar del cliente (en
     // microsegundos) en la forma de hablar del avr (en divisor de frecuencia)
     template<uint16_t periodo
-	    , uint16_t clock_frecuencia_en_hz = AVR_CLOCK_FRECUENCIA_EN_HZ>
+	    , uint32_t clock_frecuencia_en_hz = AVR_CLOCK_FRECUENCIA_EN_HZ>
     static void clock_period_in_us();
